Refactor and enhance various controllers and services
- Updated EmployeeApiController to improve pagination and error handling.
- Enhanced ProductApiController with additional error checks.
- Improved BranchController to handle missing branches gracefully.
- Refactored EmployeeController for better response handling and error management.
- Updated InventoryController to ensure consistent response types.
- Enhanced InvoiceController with improved error handling and response structure.
- Modified Auth class to return a more specific type for JWT validation.
- Improved Cache class to ensure consistent return types.
- Updated Payroll entity to ensure proper type casting.
- Enhanced Localization helper for better file handling.
- Improved decryption function in helpers for better error handling.
- Refactored Journal and Payroll repositories for better total row handling.
- Updated ProductRepository to ensure consistent naming for unit symbols.
- Enhanced HashManager to allow for string or int algorithm types.
- Refactored AccountingService and AuthService for cleaner code.
- Improved EmployeeService to handle employee creation and updates more robustly.
- Updated ReportService for better date handling.
- Refactored UserService for improved password hashing and avatar handling.
- Minor adjustments to PHPUnit configuration for clarity.
- Updated AuthServiceTest to ensure accurate password history handling.

+++ app/Entities/Payroll.php
@@ -78,7 +78,7 @@ class Payroll
 
     public function getPeriodLabel(): string
     {
-        return date('F Y', mktime(0, 0, 0, $this->month, 1, $this->year));
+        return date('F Y', (int) mktime(0, 0, 0, $this->month, 1, $this->year));
     }
 
     public function isProcessed(): bool
